+++ frontend/src/_helpers/utils.js
@@ -420,3 +420,12 @@ export function hasCircularDependency(obj) {
   }
   return false;
 }
+
+export const defineObjectProperty = (obj, key, value) =>
+  Object.defineProperties(obj, {
+    [key]: {
+      value: value,
+      writable: true,
+      enumerable: true,
+    },
+  });
